+++ src/db/Schema.php
@@ -644,38 +644,6 @@ abstract class Schema extends BaseObject
     }
 
     /**
-     * Extracts the PHP type from abstract DB type.
-     * @param ColumnSchema $column the column schema information
-     * @return string PHP type name
-     */
-    protected function getColumnPhpType($column)
-    {
-        static $typeMap = [
-            // abstract type => php type
-            self::TYPE_TINYINT => 'integer',
-            self::TYPE_SMALLINT => 'integer',
-            self::TYPE_INTEGER => 'integer',
-            self::TYPE_BIGINT => 'integer',
-            self::TYPE_BOOLEAN => 'boolean',
-            self::TYPE_FLOAT => 'double',
-            self::TYPE_DOUBLE => 'double',
-            self::TYPE_BINARY => 'resource',
-            self::TYPE_JSON => 'array',
-        ];
-        if (isset($typeMap[$column->type])) {
-            if ($column->type === 'bigint') {
-                return PHP_INT_SIZE === 8 && !$column->unsigned ? 'integer' : 'string';
-            } elseif ($column->type === 'integer') {
-                return PHP_INT_SIZE === 4 && $column->unsigned ? 'string' : 'integer';
-            }
-
-            return $typeMap[$column->type];
-        }
-
-        return 'string';
-    }
-
-    /**
      * Converts a DB exception to a more concrete one if possible.
      *
      * @param \Exception $e