+++ src/db/mssql/Schema.php
@@ -385,7 +385,7 @@ class Schema extends \yii\db\Schema implements ConstraintFinderInterface
 
         $column->type = $this->typeMap[$type] ?? self::TYPE_STRING;
 
-        $column->phpType = $this->getColumnPhpType($column);
+        $column->phpType = $column->resolvePhpType();
 
         return $column;
     }